# backend-express/docs/openapi/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Drivee Link API (Express)",
    "version": "1.0.0",
    "description": "OpenAPI path keys **không** có tiền tố `/api` vì `servers.url` trong `config/swagger.js` đã là `.../api`.\n\nKhai báo operation trong controller bằng JSDoc `@openapi` (đọc bởi `swagger-jsdoc`). Map từ Laravel: bỏ `/api` ở đầu path.\n",
    "contact": {
      "name": "ATMC Development Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3400/api",
      "description": "production"
    }
  ],
...
  "components": {
...
    "securitySchemes": {
      "auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token trả về từ POST /auth/login (field access_token, có thể đã gồm tiền tố \"Bearer \")."
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User Login",
        "operationId": "user_login",
        "description": "Laravel `@OA`: `mail_address`, `password` (query). Express đọc **JSON body** cùng tên field.\n",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mail_address",
                  "password"
                ],
                "properties": {
                  "mail_address": {
                    "type": "string",
                    "example": "1okuridashi_hanoi@gmail.vn"
                  },
                  "password": {
                    "type": "string",
                    "example": "123456789CCk"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gửi yêu cầu thành công",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": {
                    "access_token": "Bearer ...",
                    "profile": {
                      "id": 1,
                      "login_id": 111111,
                      "type": 4,
                      "status": 2,
                      "name_company_or_hro": "…",
                      "company": {
                        "jobType": {},
                        "job": {}
                      },
                      "hr_organization": {
                        "file": {},
                        "country": {}
                      },
                      "support_organization": {}
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Đăng nhập thất bại",
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "message": "Login_id or password not correct"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forget-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forget password",
        "operationId": "forget_password",
        "description": "Laravel `@OA` dùng query `mail_address`; Express dùng JSON body.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mail_address"
                ],
                "properties": {
                  "mail_address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gửi yêu cầu thành công",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": "Password reset URL has been sent to your email address"
                }
              }
            }
          }
        }
      }
    },
    "/profile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Profile",
        "operationId": "user_profile",
        "security": [
          {
            "auth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User + `name_company_or_hro` + `company` (jobType, job) / `hr_organization` (file presigned, country) / `support_organization`\n— đồng bộ cấu trúc với `AuthRepository.fillNameCompanyOrHro` sau khi cập nhật org.\n",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": {
                    "id": 1,
                    "login_id": "111111",
                    "type": 4,
                    "status": 2,
                    "mail_address": "user@example.com",
                    "parent_id": null,
                    "name_company_or_hro": "担当者1",
                    "company": {
                      "id": 1,
                      "company_name": "…",
                      "jobType": {},
                      "job": {}
                    },
                    "hr_organization": null,
                    "support_organization": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Đăng nhập thất bại",
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "message": "Sai tài khoản hoặc mật khẩu"
                }
              }
            }
          }
        }
      }
    },
    "/auth/confirm-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Confirm password",
        "operationId": "confirm_password",
        "description": "Laravel `@OA` dùng query; Express dùng JSON body `current_password`, `current_password_confirm`.",
        "security": [
          {
            "auth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "current_password",
                  "current_password_confirm"
                ],
                "properties": {
                  "current_password": {
                    "type": "string"
                  },
                  "current_password_confirm": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "send request success",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": ""
                }
              }
            }
          },
          "401": {
            "description": "password not correct",
            "content": {
              "application/json": {
                "example": {
                  "code": 401,
                  "data": "Password not correct."
                }
              }
            }
          },
          "403": {
            "description": "access denied permissions",
            "content": {
              "application/json": {
                "example": {
                  "code": 403,
                  "message": "Từ chối quyền truy cập"
                }
              }
            }
          }
        }
      }
    },
    "/auth/change-password": {
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password",
        "operationId": "change_password",
        "description": "Laravel `@OA` dùng query; Express dùng JSON body.",
        "security": [
          {
            "auth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "new_password",
                  "new_password_confirm"
                ],
                "properties": {
                  "new_password": {
                    "type": "string"
                  },
                  "new_password_confirm": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "send request success",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "data": "New password reset completed"
                }
              }
            }
          },
          "403": {
            "description": "access denied permissions",
            "content": {
              "application/json": {
                "example": {
                  "code": 403,
                  "message": "Từ chối quyền truy cập"
                }
              }
            }
          }
        }
      }
    },
...
  }
}